filesystem refactor

+++ include/xtd/mapped_file.hpp
@@ -92,7 +92,7 @@ namespace xtd{
     }
     explicit mapped_file(const filesystem::path& Path)
       : _hFile(xtd::windows::exception::throw_if(CreateFileA(Path.string().c_str(), GENERIC_READ|GENERIC_WRITE, FILE_SHARE_WRITE|FILE_SHARE_READ, nullptr, OPEN_ALWAYS, FILE_ATTRIBUTE_NORMAL, nullptr), [](HANDLE h){ return nullptr==h || INVALID_HANDLE_VALUE==h; }))
-      , _hMap(xtd::windows::exception::throw_if(CreateFileMapping(_hFile, nullptr, PAGE_READWRITE, 0, xtd::memory::page_size(), nullptr), [](HANDLE h){ return nullptr == h || INVALID_HANDLE_VALUE == h; }))      
+      , _hMap(xtd::windows::exception::throw_if(CreateFileMapping(_hFile, nullptr, PAGE_READWRITE, 0, static_cast<DWORD>(xtd::memory::page_size()), nullptr), [](HANDLE h){ return nullptr == h || INVALID_HANDLE_VALUE == h; }))      
       {}
 
     template <typename _Ty> class mapped_page{
